+++ .gitignore
@@ -1,2 +1,2 @@
-node_modules
-db
+node_modules/*
+db/*
